refactor: remove useDateTimeInput hook and extract parseDateToString and parseStringToDate functions to new files
